feat: migrate admin pages to /admin namespace
- Create Admin::BaseController for shared admin authentication
- Move UsersController to Admin::UsersController
- Update routes from /users to /admin/users
- Update Admin::ImportsController to inherit from BaseController
- Update all menu links and view paths

Closes #12
