+++ app/controllers/admin/base_controller.rb
@@ -1,0 +1,6 @@
+module Admin
+  class BaseController < ApplicationController
+    before_action :authenticate_user!
+    before_action :require_admin
+  end
+end
